Add files via upload
There is an issue with the name of some of the stair plot which display NA as the duration value, I don't know why but I m too lazy too fix it for now and it work well when I only run the individual for loop but don't work when I run the whole cell. So I don't care for now.